ka-autotest/requirements.txt
arrow==0.10.0
futures==3.0.5
beautifulsoup4==4.5.3
PyStaticConfiguration==0.10.3
toolz==0.8.2
pytest==3.2.0
pytest_allure_adaptor==1.7.8
selenium==3.5.0
pytractor==0.2.1
Pillow==4.2.1
pytesseract==0.1.7
requests==2.17.3
jinja2==2.9.5
cx-Oracle==5.1.3